+++ src/components/equipment/WeaponConfig.tsx
@@ -12,9 +12,9 @@ import { WeaponRange } from "./WeaponRangeConfig";
 import { Checkbox } from "../inputs/Checkbox";
 import { getCombatAbilities } from "../../settingsHelpers";
 import { Translate } from "../Translate";
-import { systemName } from "../../constants";
 import { assertGame } from "../../functions";
 import { assertWeaponDataSource } from "../../types";
+import { confirmADoodleDo } from "../../module/confirmADoodleDo";
 
 type WeaponConfigProps = {
   weapon: GumshoeItem,
@@ -31,35 +31,22 @@ export const WeaponConfig: React.FC<WeaponConfigProps> = ({
   const onClickDelete = useCallback(() => {
     assertGame(game);
     const message = weapon.actor
-      ? game.i18n.format(`${systemName}.DeleteActorNamesEquipmentName`, {
-          ActorName: weapon.actor.data.name,
-          EquipmentName: weapon.data.name,
-        })
-      : game.i18n.format(`${systemName}.DeleteEquipmentName`, {
-        EquipmentName: weapon.data.name,
-      });
+      ? "DeleteActorNamesEquipmentName"
+      : "DeleteEquipmentName";
 
-    const d = new Dialog({
-      title: game.i18n.localize("Confirm"),
-      content: `<p>${message}</p>`,
-      buttons: {
-        cancel: {
-          icon: '<i class="fas fa-ban"></i>',
-          label: game.i18n.localize("Cancel"),
-        },
-        delete: {
-          icon: '<i class="fas fa-trash"></i>',
-          label: game.i18n.localize("Delete"),
-          callback: () => {
-            weapon.delete();
-          },
-        },
+    confirmADoodleDo(
+      message,
+      "Delete",
+      "Cancel",
+      "fa-trash",
+      {
+        ActorName: weapon.actor?.data.name ?? "",
+        EquipmentName: weapon.data.name,
       },
-      default: "cancel",
-      // render: html => console.log("Register interactivity in the rendered dialog"),
-      // close: html => console.log("This always is logged no matter which option is chosen"),
-    });
-    d.render(true);
+      () => {
+        weapon.delete();
+      },
+    );
   }, [weapon]);
 
   const abilities = getCombatAbilities();
